Modify and create functions for results display
The main functions that impact the success of the results display are
clearResultsMessage() and addResultsMessage(). These functions allow
the outdated message from the previous round to be removed, and adds
the current round results message to display.

Other updates include: (1) change in logic for playRound() to
structure the if-else statement to divide the conditions and actions by
the results outcomes; (2) creation of capitalizeFirstLetter(); and
(3) reorder of the layout of variables and functions for improved
readability.