Create score keeper and conditions to end game
The main functions that were created are updateScore() and
getWinnerOfGame(). These functions allow the score to be updated after
each round to display who gained the point, and ends the game and
disables the buttons when either the player or the computer reaches 5
points first. These functions were added to the event listener that is
associated with the buttons.

Other updates include changes in various function and variable
names to be more descriptive.